requirements.txt
django
djangorestframework
pandas
scikit-learn